// src/lib/api/types/permissionDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface PermissionDTO {
  /** Permission code */
  code: string;
  /** Permission description */
  description?: string;
  /** Permission identifier */
  id: string;
  /** Permission module */
  module: string;
  /** Permission name */
  name: string;
}